Add files via upload
I mentioned this in the documentation, but this is not the full list. Neither the full csv or the excel file fit github's upload size limit so I had to remove what ended up being about 100,000 lines from the csv. There was most likely a much better way to work around this but I couldn't think of it.